# docs/api/simame-openapi.yaml
openapi: 3.0.3
info:
  title: Simame API
  version: 1.0.0
  description: API documentation for the Simame laundry marketplace.
paths:
...
  /api/v1/booking/lifecycle/{id}/collect-cash/:
    post:
      operationId: api_v1_booking_lifecycle_collect_cash_create
      description: Confirm COD cash received at fulfillment without touching Paystack.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
        required: true
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashCollectionResponse'
          description: ''
...
  /api/v1/booking/lifecycle/{id}/mark-picked-up/:
    patch:
      operationId: api_v1_booking_lifecycle_mark_picked_up_partial_update
      description: CONFIRMED -> PICKED_UP (Rider/Laundry)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTransition'
          description: ''
  /api/v1/booking/lifecycle/{id}/mark-washed/:
    patch:
      operationId: api_v1_booking_lifecycle_mark_washed_partial_update
      description: PICKED_UP -> IN_PROCESS (Laundry)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTransition'
          description: ''
...
  /api/v1/orders/lifecycle/{id}/collect-cash/:
    post:
      operationId: api_v1_orders_lifecycle_collect_cash_create
      description: Confirm COD cash received at fulfillment without touching Paystack.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CashCollectionRequest'
        required: true
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashCollectionResponse'
          description: ''
...
  /api/v1/orders/lifecycle/{id}/mark-picked-up/:
    patch:
      operationId: api_v1_orders_lifecycle_mark_picked_up_partial_update
      description: CONFIRMED -> PICKED_UP (Rider/Laundry)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTransition'
          description: ''
  /api/v1/orders/lifecycle/{id}/mark-washed/:
    patch:
      operationId: api_v1_orders_lifecycle_mark_washed_partial_update
      description: PICKED_UP -> IN_PROCESS (Laundry)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderTransitionRequest'
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTransition'
          description: ''
...
components:
  schemas:
...
    CashCollectionRequest:
      type: object
      properties:
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
      required:
      - amount
    CashCollectionResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        already_collected:
          type: boolean
        data:
          $ref: '#/components/schemas/OrderDetail'
      required:
      - already_collected
      - data
      - message
      - status
...
    OrderDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        order_no:
          type: string
          readOnly: true
        laundryName:
          type: string
          readOnly: true
        laundry:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatusEnum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        price_breakdown:
          type: object
          additionalProperties: {}
          readOnly: true
        pickup_date:
          type: string
          format: date-time
        delivery_date:
          type: string
          format: date-time
          nullable: true
        pickup_address:
          type: string
          nullable: true
        pickup_lat:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        pickup_lng:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        delivery_address:
          type: string
          nullable: true
        delivery_lat:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        delivery_lng:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        address:
          type: string
          nullable: true
        special_instructions:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        payment_reference:
          type: string
          nullable: true
          readOnly: true
        provider_payment_status:
          type: string
          nullable: true
          readOnly: true
        payment_method:
          $ref: '#/components/schemas/PaymentMethodBc5Enum'
        payment_state:
          type: string
          readOnly: true
        amount_due:
          type: string
          readOnly: true
        amount_collected:
          type: string
          readOnly: true
        cash_collected_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        pricing_mode:
          $ref: '#/components/schemas/PricingModeEnum'
        estimated_weight_kg:
          type: string
          format: decimal
          pattern: ^-?\d{0,4}(?:\.\d{0,2})?$
          nullable: true
      required:
      - amount_collected
      - amount_due
      - cash_collected_at
      - created_at
      - id
      - items
      - laundry
      - laundryName
      - order_no
      - payment_reference
      - payment_state
      - pickup_date
      - price_breakdown
      - provider_payment_status
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        item:
          type: string
          format: uuid
          nullable: true
        service_type:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          readOnly: true
        quantity:
          type: integer
          maximum: 2147483647
          minimum: 0
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - id
      - name
      - price
...
    OrderStatusEnum:
      enum:
      - PENDING
      - CONFIRMED
      - REJECTED
      - PICKED_UP
      - IN_PROCESS
      - OUT_FOR_DELIVERY
      - DELIVERED
      - COMPLETED
      - CANCELLED
      type: string
      description: |-
        * `PENDING` - Pending
        * `CONFIRMED` - Confirmed
        * `REJECTED` - Rejected
        * `PICKED_UP` - Picked Up
        * `IN_PROCESS` - In Process
        * `OUT_FOR_DELIVERY` - Out for Delivery
        * `DELIVERED` - Delivered
        * `COMPLETED` - Completed
        * `CANCELLED` - Cancelled
    OrderTransition:
      type: object
      description: Generic serializer for status transitions that may require a reason.
      properties:
        reason:
          type: string
        metadata: {}
...
    PatchedOrderTransitionRequest:
      type: object
      description: Generic serializer for status transitions that may require a reason.
      properties:
        reason:
          type: string
        metadata: {}
...
    PaymentMethodBc5Enum:
      enum:
      - CARD
      - BANK_TRANSFER
      - CASH
      type: string
      description: |-
        * `CARD` - Paystack
        * `BANK_TRANSFER` - Bank transfer
        * `CASH` - Cash on delivery
...
    PaymentStatusEnum:
      enum:
      - PAID
      - UNPAID
      - REFUNDED
      type: string
      description: |-
        * `PAID` - Paid
        * `UNPAID` - Unpaid
        * `REFUNDED` - Refunded
...
    PricingModeEnum:
      enum:
      - BY_ITEM
      - BY_WEIGHT
      - CUSTOM_QUOTE
      type: string
      description: |-
        * `BY_ITEM` - By item
        * `BY_WEIGHT` - By weight
        * `CUSTOM_QUOTE` - Pay after quote
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Connect SimpleJWT access token or verified Clerk session JWT.
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid